% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bathy-data.R, R/SOmap_data.R
\docType{data}
\name{Bathy}
\alias{Bathy}
\title{Bathymetric data for maps}
\format{An object of class \code{"RasterLayer"}}
\source{
\href{https://www.gebco.net/data_and_products/gridded_bathymetry_data/}{GEBCO}

\href{https://www.gebco.net/data_and_products/gridded_bathymetry_data/}{GEBCO}
}
\usage{
data(Bathy)

data(Bathy)
}
\description{
Bathymetric data reprocessed from the GEBCO_2014 Grid data set.

Bathymetric data reprocessed from the GEBCO_2014 Grid data set.
}
\references{
The GEBCO_2014 Grid, version 20150318

The GEBCO_2014 Grid, version 20150318
}
\keyword{datasets}
