% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{coord_sf_unicode}
\alias{coord_sf_unicode}
\alias{geom_sf_unicode}
\title{Replacement ggplot functions}
\usage{
coord_sf_unicode(...)

geom_sf_unicode(...)
}
\arguments{
\item{...}{: as for \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}} or \code{\link[ggplot2:ggsf]{ggplot2::coord_sf()}}}
}
\description{
There is an issue with the default graphics device under Windows and R 4.2.2, which prevents degree symbols being shown correctly (see \url{https://github.com/tidyverse/ggplot2/issues/5147} and \url{https://github.com/r-lib/ragg/issues/51}). This function provides a (hopefully temporary) workaround by using the Unicode degree symbol into the labels.
}
