% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_SOmap.R
\name{default_somap}
\alias{default_somap}
\title{Default southern ocean map}
\usage{
default_somap(xs, ys, centre_lon = NULL, centre_lat = NULL,
  family = "stere", dimXY = c(300, 300), bathy = TRUE,
  coast = TRUE, input_points = TRUE, input_lines = TRUE,
  graticule = TRUE)
}
\arguments{
\item{xs, ys}{longitude and latitude values}

\item{centre_lon, centre_lat}{optional map centre longitude and latitude}

\item{family}{optional projection family (default is `stere`ographic)}

\item{dimXY}{dimensions of background bathmetry (if used) default is 300x300}

\item{bathy, coast}{optional input bathymetry or coastline data, defaults to `TRUE` set to `FALSE` to ignore or input your own}

\item{input_points, input_lines}{flag to plot input data as points and / or lines}

\item{graticule}{flag to add a basic graticule}
}
\value{
the derived target extent in the map projection used, bathymetry, and coastline data
}
\description{
Provide minimal input information to get a default map. The simplest case is
to run the function without any inputs at all and it will provide a random default.
}
\details{
To input your data, use input locations as `xs` (longitude) and `ys` (latitude) values, there must be at least
two locations.

Try families such as `laea`, `ortho`, `gnomonic` if feeling adventurous.
}
\examples{
default_somap(c(0, 50), c(-70, -50))
default_somap(runif(10, 130, 200), runif(10, -80, -10))
default_somap(runif(10, 130, 200), runif(10, -85, -60))
## save the result to explore later!
protomap <- default_somap(runif(10, 60, 160), runif(10, -73, -50), coast = rnaturalearth::ne_coastline())
\dontrun{
library(lazyraster)
ibcso <- as_raster(lazyraster(raadtools::topofile("ibcso")), dim = c(2000, 2000))
projection(ibcso) <- projection(raadtools::readtopo("ibcso"))
default_somap(runif(10, 30, 160), runif(10, -75, -40), bathy = ibcso)
}
}
