% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{SO_plotter}
\alias{SO_plotter}
\title{Construct a SO_plotter object}
\usage{
SO_plotter(plotfun, plotargs = NULL, name = NULL)
}
\arguments{
\item{plotfun}{function or string: either the name of a function to use, or the function itself}

\item{plotargs}{list: arguments to pass to the function}

\item{name}{string: optional name for this element}
}
\value{
An object of class SO_plotter
}
\description{
\code{SOmap} and similar objects contain all of the data and code required to draw a map. This information is embedded in \code{SO_plotter} objects within the \code{SOmap} object.
}
\examples{
\dontrun{
  p <- SOmap()
  ## replace the `box` element with different plotting code
  p$box <- SO_plotter(plotfun = "graphics::box", plotargs = list(col = "red"))

  ## you can also specify multiple plotting instructions for a single graphical element
  ##  of a map
  p$box <- c(SO_plotter(plotfun = "graphics::box", plotargs = list(col = "red")),
             SO_plotter(plotfun = "graphics::box", plotargs = list(lwd = 2)))
}
}
\seealso{
\code{SOmap}
}
