% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOauto_crop.r
\name{SOauto_crop}
\alias{SOauto_crop}
\title{Reproject and crop Spatial Polygons to SOmap objects}
\usage{
SOauto_crop(layer, x)
}
\arguments{
\item{layer}{A SpatialPolygonsDataFrame.}

\item{x}{A SOmap or SOauto_map object.}
}
\value{
reprojected and cropped version of layer
}
\description{
Reproject and crop Spatial Polygons to SOmap objects
}
\examples{
\dontrun{
a<-SOauto_map(c(0, 50), c(-70, -50))
x<-SOauto_crop(SOmap_data$fronts_orsi,  a)
a
plot(x, add = TRUE)
}

}
