% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOcode.R
\name{SOcode}
\alias{SOcode}
\title{Extract plotting code from a SOmap}
\usage{
SOcode(x, data_object_name = "SOmap_data")
}
\arguments{
\item{x}{: a map object as returned by \link{SOmap}, \link{SOmanagement}, \link{SOleg}, or \link{SOgg}}

\item{data_object_name}{string: the name to use for the object that will hold the map data. See Examples, below}
}
\value{
A list with two elements: \code{code} contains R code that will draw the map, and \code{SOmap_data} (or whatever was passed as the \code{data_object_name} argument) contains any data required by that code
}
\description{
This is thoroughly experimental!
}
\examples{
\dontrun{
  p <- SOmap()
  mapcode <- SOcode(p, data_object_name = "SOmap_data")

  ## write this code to a file
  my_R_file <- tempfile(fileext = ".R")
  writeLines(mapcode$code, con = my_R_file)

  ## you can edit the code in that file if desired

  ## save the data
  my_data_file <- tempfile(fileext = ".rds")
  saveRDS(mapcode$SOmap_data, my_data_file)

  ## later on, we can re-load the data and execute the code
  SOmap_data <- readRDS(my_data_file)
  source(my_R_file)

  ## or just to show that this works, evaluate the returned code directly against its data
  with(mapcode, for (codeline in code) eval(parse(text = codeline)))
}

}
\seealso{
\link{SOmap}
}
