% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{SOgg_cex}
\alias{SOgg_cex}
\title{Convert cex to ggplot size}
\usage{
SOgg_cex(cex)
}
\arguments{
\item{cex}{numeric: character expansion, see \code{\link{par}}}
}
\value{
The corresponding 'size' value to use in ggplot calls
}
\description{
Text size in base graphics is generally specified via \code{cex} values, which are multipliers applied to the device pointsize. \code{SOgg_cex} is a convenience function that converts a cex value into a \code{size} value as used by ggplot2 geometries.
}
