% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOleg.R
\name{SOleg}
\alias{SOleg}
\title{SOleg creating rounded legends for SOmap}
\usage{
SOleg(x, position = "topright", col = grDevices::heat.colors(80),
  ticks = 6, tlabs = "", Trim = -45, label = "Heat")
}
\arguments{
\item{x}{Object to plot (currently not used)}

\item{position}{Where you want the legend ("topleft","topright", "bottomright")}

\item{col}{Color pattern to use.}

\item{ticks}{Number of ticks to include on the legend.}

\item{tlabs}{Tick labels.}

\item{Trim}{Trim that was used to create the SOmap.}

\item{label}{Legend label.}
}
\value{
Creates rounded legends
}
\description{
SOleg creating rounded legends for SOmap
}
\examples{
\dontrun{
SOleg(position="topleft",
      col=heat.colors(80),
      ticks=4,
      tlabs = c("1","10","100","1000"),
      Trim=-45,
      label="Heat")
}
}
