% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOmap_auto.R
\name{SOmap-defunct}
\alias{SOmap-defunct}
\alias{default_somap}
\alias{SOauto_map}
\title{Defunct function}
\usage{
default_somap(...)

SOauto_map(...)
}
\arguments{
\item{...}{all arguments passed to new function}
}
\description{
Removed from SOmap
}
