% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOmap.R
\name{SOmap}
\alias{SOmap}
\title{Southern Ocean Map}
\usage{
SOmap(Bathleg = TRUE, Border = TRUE, Trim = -45, Grats = FALSE,
  straight = FALSE, land = TRUE, fronts = FALSE,
  frontcols = c("hotpink", "orchid", "plum"), bordercol = c("white",
  "black"), gratcol = "grey70")
}
\arguments{
\item{Bathleg}{Insert the bathymetry legend.}

\item{Border}{Insert longitude border.}

\item{Trim}{Longitude to trim map to.}

\item{Grats}{Insert graticule grid.}

\item{straight}{Do you need a blank space on the side for a straight legend.}

\item{land}{Plot land boundary}

\item{fronts}{Plot ocean fronts: Subantarctic Front, Polar Front, Southern Antarctic Circumpolar Current Front}

\item{frontcols}{colors for fronts

 @return
Produces at the very base a round bathymetry map of the southern hemisphere.}

\item{bordercol}{Colors for longitude border; Default is c("black","white").}

\item{gratcol}{Color for graticule grid; Default is grey.}
}
\description{
Function for creating round Southern Ocean maps.
}
\examples{
Dat.Dir<-getwd()
png(paste(Dat.Dir,'/SOmap.png', sep=''), width=22, height=20, units='cm', res=600)
SOmap(Trim=-45, grats=T)
dev.off()

}
