% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOmap2.R
\name{SOmap2}
\alias{SOmap2}
\title{Southern Ocean Map 2}
\usage{
SOmap2(bathy_legend = TRUE, land = TRUE, ice = TRUE,
  ccamlr = FALSE, ccamlr_labels = FALSE, ssru = FALSE,
  ssru_labels = FALSE, ssmu = FALSE, ssmu_labels = FALSE,
  rb = FALSE, rb_labels = FALSE, sprfmorb = FALSE, border = TRUE,
  trim = -45, graticules = FALSE, eez = FALSE, eez_labels = FALSE,
  mpa = FALSE, mpa_labels = FALSE, domains = FALSE,
  domains_labels = FALSE, iwc = FALSE, iwc_labels = FALSE,
  straight = FALSE, fronts = FALSE, fronts_col = c("hotpink",
  "orchid", "plum"), land_col = "black", ice_col = "black",
  rb_col = 3, sprfmo_col = "grey50", ccamlr_col = 2,
  ssru_col = "grey50", ssmu_col = "grey70", eez_col = "maroon",
  mpa_col = "yellow", border_col = c("white", "black"),
  graticules_col = "grey70", iwc_col = "blue",
  domains_col = "magenta")
}
\arguments{
\item{bathy_legend}{Insert the bathymetry legend.}

\item{land}{Plot land boundary}

\item{ice}{logical: if \code{TRUE}, plot ice features (ice shelves, glacier tongues, and similar)}

\item{ccamlr}{Insert the CCAMLR boundaries.}

\item{ccamlr_labels}{Insert the CCAMLR labels.}

\item{ssru}{Insert the CCAMLR small scale research unit boundaries.}

\item{ssru_labels}{Insert the CCAMLR small scale research unit labels.}

\item{ssmu}{Insert the CCAMLR small scale management unit boundaries.}

\item{ssmu_labels}{Insert the CCAMLR small scale management unit labels.}

\item{rb}{Insert the CCAMLR research block boundaries.}

\item{rb_labels}{Insert the CCAMLR research block labels.}

\item{sprfmorb}{Insert the SPRFMO toothfish research block boundaries.}

\item{border}{Insert longitude border.}

\item{trim}{Longitude to trim map to.}

\item{graticules}{Insert graticule grid.}

\item{eez}{Insert Exclusive Economic Zones.}

\item{eez_labels}{Insert Exclusive Economic Zone labels.}

\item{mpa}{Insert CCAMLR Marine Protected Areas.}

\item{mpa_labels}{Insert CCAMLR Marine Protected Area labels.}

\item{domains}{Insert CCAMLR Marine Protected Areas planning domains.}

\item{domains_labels}{Insert CCAMLR Marine Protected Area planning domains labels.}

\item{iwc}{Insert International Whaling Commission boundaries.}

\item{iwc_labels}{Insert International Whaling Commission labels.}

\item{straight}{Do you need a blank space on the side for a straight legend.}

\item{fronts}{Plot ocean fronts: Subantarctic Front, Polar Front, Southern Antarctic Circumpolar Current Front}

\item{fronts_col}{colors for fronts}

\item{land_col}{character: colour to use for coastline}

\item{ice_col}{character: colour to use for ice features}

\item{rb_col}{Color for CCAMLR research blocks.}

\item{sprfmo_col}{Color for SPRFMO toothfish research blocks}

\item{ccamlr_col}{Color for CCAMLR boundaries}

\item{ssru_col}{Color for CCAMLR small scale research units.}

\item{ssmu_col}{Color for CCAMLR small scale management units.}

\item{eez_col}{Color for Exclusive Economic Zone boundaries; Default is maroon.}

\item{mpa_col}{Color for CCAMLR Marine Protected Areas; Default is yellow.}

\item{border_col}{Colors for longitude border; Default is c("black","white").}

\item{graticules_col}{Color for graticule grid; Default is grey.}

\item{iwc_col}{Color for IWC boundaries; Default is blue.}

\item{domains_col}{Color for the CCAMLR planning domains boundaries. Default is magenta.}
}
\value{
An object of class "SOmap", which represents a polar-stereographic map of the southern hemisphere, with the chosen management layers added. Printing or plotting this object will cause it to be displayed in the current graphics device.
}
\description{
Function for creating round Southern Ocean maps with inbuild base layers.
}
\examples{
\dontrun{
  SOmap2(ccamlr = TRUE, mpa = TRUE, trim = -45)
}
}
