% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOmap_data.R
\docType{data}
\name{SOmap_data}
\alias{SOmap_data}
\title{Contextual data for Southern Ocean maps}
\format{A list containing the following elements:
\itemize{
  \item CCAMLR_MPA
    \itemize{
      \item Description: current marine protected areas
      \item Source: CCAMLR
      \item URL: https://data.ccamlr.org/dataset/marine-protected-areas
      \item License: not specified
    }
  \item CCAMLR_statistical_areas
    \itemize{
      \item Description: Statistical areas, subareas and divisions are used globally for the purpose of reporting fishery statistics. CCAMLR's Convention Area in the Southern Ocean is divided, for statistical purposes, into Area 48 (Atlantic Antarctic) between 70W and 30E, Area 58 (Indian Ocean Antarctic) between 30 and 150E, and Area 88 (Pacific Antarctic) between 150E and 70W. These areas, which are further subdivided into subareas and divisions, are managed by CCAMLR.
      \item Source: CCAMLR
      \item URL: https://data.ccamlr.org/dataset/statistical-areas-subareas-and-divisions
      \item License: Public domain
    }
}}
\usage{
data(SOmap_data)
}
\description{
Various spatial datasets that are commonly used on Southern Ocean maps.
}
\keyword{datasets}
