% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOmanagement.R
\name{SOmap_text}
\alias{SOmap_text}
\title{Helper function for labels
This is basically a thin wrapper around \code{text}, that passes \code{x[[labelcol]]} to \code{text} as the \code{labels} parameter}
\usage{
SOmap_text(x, labelcol, ...)
}
\arguments{
\item{x}{data.frame or Spatial data.frame: data to pass to \code{text}}

\item{labelcol}{string: name of the column in \code{x} to use for text labels}

\item{...}{other plot arguements}
}
\value{
as for \code{text}
}
\description{
Helper function for labels
This is basically a thin wrapper around \code{text}, that passes \code{x[[labelcol]]} to \code{text} as the \code{labels} parameter
}
\seealso{
\code{\link{text}}
}
