% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOproj.R
\name{SOproj}
\alias{SOproj}
\title{Southern projection}
\usage{
SOproj(x, y = NULL, target = NULL, data, ..., source = NULL)
}
\arguments{
\item{x}{longitude vector, or object with coordinates}

\item{y}{latitude vector}

\item{target}{target projection (default = stereo)}

\item{data}{optional data to be included}

\item{...}{arguments passed to \code{\link[reproj:reproj]{reproj::reproj()}}}

\item{source}{starting projection (default = longlat)}
}
\value{
Reprojects the given data object to polar projection. Works with Points, spatial, raster, SOmap, sf and sfc objects.
}
\description{
Function for reprojecting data.
}
\examples{
\dontrun{
 lat <- c(-70, -60,-50, -90)
 lon <- c(-50, -75, -45, -60)
 pnts <- SOproj(x = lon, y = lat)
 SOmap2(CCAMLR = TRUE)
 plot(pnts, pch = 19, col = 3, add = TRUE)
}
}
