% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_SOmap.R
\name{default_somap}
\alias{default_somap}
\title{Default southern ocean map}
\usage{
default_somap(xs, ys, centre_lon = NULL, centre_lat = NULL,
  family = "stere", expand = TRUE, dimXY = c(300, 300),
  bathy = TRUE, coast = TRUE, input_points = TRUE,
  input_lines = TRUE, graticule = TRUE, buffer = 0.05,
  contours = TRUE, lvs = c(-500, -1000, -2000))
}
\arguments{
\item{xs}{optional input data longitudes}

\item{ys}{optional input data latitudes}

\item{centre_lon}{optional centre longitude (of the map projection, also used to for plot range if `expand = TRUE`)}

\item{centre_lat}{as per `centre_lon`}

\item{family}{optional projection family (default is `stere`ographic)}

\item{expand}{re-compute range of plot to incorporate centre_lon and centre_lat with the data as a natural middle}

\item{dimXY}{dimensions of background bathmetry (if used) default is 300x300}

\item{bathy}{optional bathymetry data to use (or `FALSE` for no bathmetry image)}

\item{coast}{optional coastline data to use (or `FALSE` for no coastline)}

\item{input_points}{add points to plot (of xs, ys)}

\item{input_lines}{add lines to plot   (of xs, ys)}

\item{graticule}{flag to add a basic graticule}

\item{buffer}{fraction to expand plot range from that calculated (either from data, or from centre_lon/centre_lat _and_ data if `expand = TRUE`)}

\item{contours}{add contours}

\item{lvs}{contour levels if `contours = TRUE`}
}
\value{
the derived target extent and the map projection used, bathymetry, and coastline data
}
\description{
Provide minimal input information to get a default map. The simplest case is
to run the function without any inputs at all and it will provide a random default.
}
\details{
To input your data, use input locations as `xs` (longitude) and `ys` (latitude) values, there must be at least
two locations.

Try families such as `laea`, `ortho`, `gnomonic` if feeling adventurous.
}
\examples{
default_somap(c(0, 50), c(-70, -50))
default_somap(runif(10, 130, 200), runif(10, -80, -10))
default_somap(runif(10, 130, 200), runif(10, -85, -60))
## save the result to explore later!
protomap <- default_somap(runif(10, 60, 160), runif(10, -73, -50))
}
