% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SObin.R
\name{SObin}
\alias{SObin}
\title{Bin longitude latitude points by count in a SOmap context}
\usage{
SObin(
  x,
  y = NULL,
  baselayer = NULL,
  ...,
  col = hcl.colors(26, "Viridis"),
  dim = c(512, 512),
  add = TRUE,
  target = NULL,
  source = NULL,
  data.frame = FALSE
)
}
\arguments{
\item{x}{longitudes}

\item{y}{latitudes}

\item{baselayer}{optional spatial layer to get extent from}

\item{...}{passed to plot if \code{add = TRUE}}

\item{col}{colours to use if \code{add = TRUE}}

\item{dim}{dimensions of raster to bin to}

\item{add}{if \code{TRUE}, the raster is added to the current plot. An error is thrown if there is no existing plot}

\item{target}{target projection passed to SOproj}

\item{source}{source projection of data projection passed to SOproj}

\item{data.frame}{if true return a data frame instead of a raster.}
}
\value{
A raster. If \code{add = TRUE}, it is returned invisibly.
}
\description{
Creates a raster density layer from a set of points.
}
\examples{
SOmap_auto()
pts <- geosphere::randomCoordinates(1e6)
bin <- SObin(pts[, 1], pts[, 2], add = TRUE)
}
