% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{SOgg}
\alias{SOgg}
\title{Generate a ggplot2 representation of an SOmap object}
\usage{
SOgg(x)
}
\arguments{
\item{x}{SOmap or SOmap_auto: object as returned by \code{SOmap}, \code{SOmap2}, or \code{SOmap_auto}}
}
\value{
A ggplot object
}
\description{
Note: this function is still experimental! Use at your own risk.
}
\examples{
\dontrun{
  p <- SOmap2(trim = -45, iwc = TRUE, iwc_labels = TRUE, graticules = TRUE, fronts = TRUE,
              mpa = TRUE, mpa_labels = TRUE)
  SOgg(p)
}

}
