% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOleg.R
\name{SOleg}
\alias{SOleg}
\title{Rounded legends for SOmap}
\usage{
SOleg(x = NULL, position = "topright", col = NULL, ticks = NULL,
  tlabs = NULL, breaks = NULL, trim = -45, type = "discrete",
  label = "", ladj = 0.5, lsrt = 0, lcex = 0.75, tadj = 0.5,
  tcex = 1, rnd = NULL, border_width = 2)
}
\arguments{
\item{x}{numeric: object to obtain min and max values from for \code{type = "continuous"}.}

\item{position}{string: where you want the legend ("topleft", "topright", "bottomleft", or "bottomright").}

\item{col}{character: colours to use.}

\item{ticks}{numeric: number of ticks to include on the legend. Only used with \code{type = "continuous"}.}

\item{tlabs}{character: tick labels. Required for \code{type = "discrete"}, optional for \code{type = "continuous"} if \code{x} is given.}

\item{breaks}{numeric: vector of tick positions for \code{type = "continuous"} when \code{x} is given.}

\item{trim}{numeric: \code{trim} value that was used to create the SOmap object (see \code{\link{SOmap}}).}

\item{type}{string: type of legend ("discrete" or "continuous").}

\item{label}{string: legend label.}

\item{ladj}{numeric: distance to adjust the tick labels from the ticks.}

\item{lsrt}{numeric: angle of the tick labels.}

\item{lcex}{numeric: size of the tick labels.}

\item{tadj}{numeric: distance to adjust the title from the ticks.}

\item{tcex}{numeric: size of the title text.}

\item{rnd}{numeric: optional rounding factor for continuous legends using the \code{link{round}} function.}

\item{border_width}{numeric: thickness (in degrees of latitude) of the border.}
}
\value{
An object of class "SOmap_legend". Printing or plotting this object will cause it to be added to the SOmap in the current graphics device.
}
\description{
Rounded legends for SOmap
}
\examples{
\dontrun{
  SOmap()

  ## Discrete Legend
  SOleg(position = "topleft", col = viridis::viridis(5),
        tlabs = c("a", "b", "c", "d", "e"), trim = -45, label = "Species")

  ## Continuous Legend
  SOleg(x = runif(100), position = "topright", col = viridis::viridis(80),
        breaks = c(0.1, 0.2, 0.5, 0.9), trim = -45, label = "Species",
        rnd = 1, type = "continuous")
}
}
