% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOmanagement.R
\name{SOmanagement}
\alias{SOmanagement}
\title{Southern Ocean management map layers}
\usage{
SOmanagement(ccamlr = FALSE, ccamlr_labels = FALSE, ssru = FALSE,
  ssru_labels = FALSE, ssmu = FALSE, ssmu_labels = FALSE,
  rb = FALSE, rb_labels = FALSE, sprfmorb = FALSE, trim = -45,
  eez = FALSE, eez_labels = FALSE, mpa = FALSE, mpa_labels = FALSE,
  iwc = FALSE, iwc_labels = FALSE, domains = FALSE,
  domains_labels = FALSE, rb_col = "green", sprfmo_col = "grey50",
  ccamlr_col = "red", ssru_col = "grey50", ssmu_col = "grey70",
  eez_col = "maroon", mpa_col = "yellow", iwc_col = "blue",
  domains_col = "magenta", basemap)
}
\arguments{
\item{ccamlr}{Insert the CCAMLR boundaries.}

\item{ccamlr_labels}{Insert the CCAMLR labels.}

\item{ssru}{Insert the CCAMLR small scale research unit boundaries.}

\item{ssru_labels}{Insert the CCAMLR small scale research unit labels.}

\item{ssmu}{Insert the CCAMLR small scale management unit boundaries.}

\item{ssmu_labels}{Insert the CCAMLR small scale management unit labels.}

\item{rb}{Insert the CCAMLR research block boundaries.}

\item{rb_labels}{Insert the CCAMLR research block labels.}

\item{sprfmorb}{Insert the SPRFMO toothfish research block boundaries.}

\item{trim}{Longitude to trim map to.}

\item{eez}{Insert Exclusive Economic Zones.}

\item{eez_labels}{Insert Exclusive Economic Zone labels.}

\item{mpa}{Insert CCAMLR Marine Protected Areas.}

\item{mpa_labels}{Insert CCAMLR Marine Protected Area labels.}

\item{iwc}{Insert International Whaling Commission boundaries.}

\item{iwc_labels}{Insert International Whaling Commission labels.}

\item{domains}{Insert CCAMLR Marine Protected Area planning domains.}

\item{domains_labels}{Insert CCAMLR Marine Protected Area planning domains labels.}

\item{rb_col}{Color for CCAMLR research blocks.}

\item{sprfmo_col}{Color for SPRFMO toothfish research blocks}

\item{ccamlr_col}{Color for CCAMLR boundaries}

\item{ssru_col}{Color for CCAMLR small scale research units.}

\item{ssmu_col}{Color for CCAMLR small scale management units.}

\item{eez_col}{Color for Exclusive Economic Zone boundaries; Default is maroon.}

\item{mpa_col}{Color for CCAMLR Marine Protected Areas; Default is yellow.}

\item{iwc_col}{Color for IWC boundaries; Default is blue.}

\item{domains_col}{Color for the CCAMLR planning domains boundaries. Default is magenta.}

\item{basemap}{Optional SOmap or SOmap_auto object.}
}
\value{
An object of class "SOmap_management" containing the requested management layers. Printing or plotting this object will display those layers on the current map (note that an \code{SOmap} object needs to have been plotted first)
}
\description{
Function for adding management layers to SOmap
}
\examples{
\dontrun{
  tfile <- tempfile("SOmap", fileext = ".png")
  png(tfile, width=22, height=20, units='cm', res=600)
  SOmap(trim = -45)
  SOmanagement(ccamlr = TRUE, ccamlr_labels = TRUE, trim=-45)
  dev.off()
  unlink(tfile)

  SOmap(trim = -45)
  SOmanagement(ccamlr = TRUE, ccamlr_labels = TRUE, trim = -45)
}
}
