% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOmanagement.R
\name{SOmanagement}
\alias{SOmanagement}
\title{Southern Ocean management map layers}
\usage{
SOmanagement(ccamlr = FALSE, ccamlr_labels = FALSE, ssru = FALSE,
  ssru_labels = FALSE, ssmu = FALSE, ssmu_labels = FALSE,
  rb = FALSE, rb_labels = FALSE, sprfmorb = FALSE, trim = -45,
  eez = FALSE, eez_labels = FALSE, mpa = FALSE, mpa_labels = FALSE,
  iwc = FALSE, iwc_labels = FALSE, domains = FALSE,
  domains_labels = FALSE, rb_col = "green", sprfmo_col = "grey50",
  ccamlr_col = "red", ssru_col = "grey50", ssmu_col = "grey70",
  eez_col = "maroon", mpa_col = "yellow", iwc_col = "blue",
  domains_col = "magenta", basemap)
}
\arguments{
\item{ccamlr}{logical: if \code{TRUE}, insert the CCAMLR area boundaries.}

\item{ccamlr_labels}{logical: if \code{TRUE}, add labels for the CCAMLR areas.}

\item{ssru}{logical: if \code{TRUE}, insert the CCAMLR small scale research unit boundaries.}

\item{ssru_labels}{logical: if \code{TRUE}, add labels for the CCAMLR small scale research units.}

\item{ssmu}{logical: if \code{TRUE}, insert the CCAMLR small scale management unit boundaries.}

\item{ssmu_labels}{logical: if \code{TRUE}, add labels for the CCAMLR small scale management units.}

\item{rb}{logical: if \code{TRUE}, insert the CCAMLR research block boundaries.}

\item{rb_labels}{logical: if \code{TRUE}, add labels for the CCAMLR research blocks.}

\item{sprfmorb}{logical: if \code{TRUE}, insert the SPRFMO toothfish research block boundaries.}

\item{trim}{numeric: latitude to trim the map to. Set this to -10 for effectively no trim.}

\item{eez}{logical: if \code{TRUE}, insert Exclusive Economic Zones.}

\item{eez_labels}{logical: if \code{TRUE}, add labels for the Exclusive Economic Zones.}

\item{mpa}{logical: if \code{TRUE}, insert CCAMLR Marine Protected Areas.}

\item{mpa_labels}{logical: if \code{TRUE}, add labels for the CCAMLR Marine Protected Areas.}

\item{iwc}{logical: if \code{TRUE}, insert International Whaling Commission boundaries.}

\item{iwc_labels}{logical: if \code{TRUE}, add labels for the International Whaling Commission areas.}

\item{domains}{logical: if \code{TRUE}, insert CCAMLR Marine Protected Areas planning domains.}

\item{domains_labels}{logical: if \code{TRUE}, add labels for the CCAMLR Marine Protected Area planning domains.}

\item{rb_col}{character: colour for CCAMLR research blocks.}

\item{sprfmo_col}{character: colour for SPRFMO toothfish research blocks}

\item{ccamlr_col}{character: colour for CCAMLR boundaries}

\item{ssru_col}{character: colour for CCAMLR small scale research units.}

\item{ssmu_col}{character: colour for CCAMLR small scale management units.}

\item{eez_col}{character: colour for Exclusive Economic Zone boundaries.}

\item{mpa_col}{character: colour for CCAMLR Marine Protected Areas.}

\item{iwc_col}{character: colour for IWC boundaries.}

\item{domains_col}{character: colour for the CCAMLR planning domains boundaries.}

\item{basemap}{SOmap or SOmap_auto: optional map object to extract extent, projection, and other information from.}
}
\value{
An object of class "SOmap_management" containing the requested management layers. Printing or plotting this object will display those layers on the current map (note that an \code{SOmap} object needs to have been plotted first)
}
\description{
Function for adding management layers to SOmap
}
\examples{
\dontrun{
  tfile <- tempfile("SOmap", fileext = ".png")
  png(tfile, width=22, height=20, units='cm', res=600)
  SOmap(trim = -45)
  SOmanagement(ccamlr = TRUE, ccamlr_labels = TRUE, trim=-45)
  dev.off()
  unlink(tfile)

  SOmap(trim = -45)
  SOmanagement(ccamlr = TRUE, ccamlr_labels = TRUE, trim = -45)
}
}
