% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOmap-package.R
\docType{package}
\name{SOmap-package}
\alias{SOmap-package}
\title{\pkg{SOmap}}
\description{
Create publication-quality Southern Ocean maps in a simple manner with multiple management layer options.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AustralianAntarcticDivision/SOmap}
  \item Report bugs at \url{https://github.com/AustralianAntarcticDivision/SOmap/issues}
}

}
\author{
\strong{Maintainer}: Dale Maschette \email{dale.maschette@aad.gov.au}

Authors:
\itemize{
  \item Michael Sumner
  \item Ben Raymond
}

}
\keyword{internal}
