% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOmap.R
\name{SOmap}
\alias{SOmap}
\title{Southern Ocean Map}
\usage{
SOmap(
  bathy_legend = TRUE,
  border = TRUE,
  trim = -45,
  graticules = FALSE,
  straight = FALSE,
  land = TRUE,
  land_col = "black",
  ice = TRUE,
  ice_col = "black",
  fronts = FALSE,
  fronts_col = c("hotpink", "orchid", "plum"),
  border_col = c("white", "black"),
  border_width = 2,
  graticules_col = "grey70"
)
}
\arguments{
\item{bathy_legend}{logical: if \code{TRUE}, insert the bathymetry legend.}

\item{border}{logical: if \code{TRUE}, insert longitude border.}

\item{trim}{numeric: latitude to trim the map to. Set this to -10 for effectively no trim.}

\item{graticules}{logical: if \code{TRUE}, insert graticule grid.}

\item{straight}{logical: if \code{TRUE}, leave a blank space on the side for a straight legend.}

\item{land}{logical: if \code{TRUE}, plot coastline.}

\item{land_col}{character: colour to use for coastline.}

\item{ice}{logical: if \code{TRUE}, plot ice features (ice shelves, glacier tongues, and similar).}

\item{ice_col}{character: colour to use for ice features.}

\item{fronts}{logical: if \code{TRUE}, plot ocean fronts (Subantarctic Front, Polar Front, Southern Antarctic Circumpolar Current Front).}

\item{fronts_col}{character: colours for fronts.}

\item{border_col}{character: colours for longitude border.}

\item{border_width}{numeric: thickness (in degrees of latitude) of the border.}

\item{graticules_col}{string: colour for graticule grid.}
}
\value{
An object of class "SOmap", which represents a polar-stereographic map of the southern hemisphere. Printing or plotting this object will cause it to be displayed in the current graphics device.
}
\description{
Function for creating round Southern Ocean maps.
}
\examples{
\dontrun{
  tfile <- tempfile("SOmap", fileext = ".png")
  png(tfile, width = 22, height = 20, units = "cm", res = 600)
  SOmap(trim = -45, graticules = TRUE)
  dev.off()
  unlink(tfile)
  SOmap(trim = -45, graticules = TRUE)
}
}
