% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOmap2.R
\name{SOmap2}
\alias{SOmap2}
\title{Southern Ocean Map 2}
\usage{
SOmap2(
  bathy_legend = TRUE,
  land = TRUE,
  ice = TRUE,
  ccamlr = FALSE,
  ccamlr_labels = FALSE,
  ssru = FALSE,
  ssru_labels = FALSE,
  ssmu = FALSE,
  ssmu_labels = FALSE,
  rb = FALSE,
  rb_labels = FALSE,
  sprfmorb = FALSE,
  border = TRUE,
  trim = -45,
  graticules = FALSE,
  eez = FALSE,
  eez_labels = FALSE,
  mpa = FALSE,
  mpa_labels = FALSE,
  domains = FALSE,
  domains_labels = FALSE,
  iwc = FALSE,
  iwc_labels = FALSE,
  straight = FALSE,
  fronts = FALSE,
  fronts_col = c("hotpink", "orchid", "plum"),
  land_col = "black",
  ice_col = "black",
  rb_col = 3,
  sprfmo_col = "grey50",
  ccamlr_col = 2,
  ssru_col = "grey50",
  ssmu_col = "grey70",
  eez_col = "maroon",
  mpa_col = "yellow",
  border_col = c("white", "black"),
  graticules_col = "grey70",
  iwc_col = "blue",
  domains_col = "magenta"
)
}
\arguments{
\item{bathy_legend}{logical: if \code{TRUE}, insert the bathymetry legend.}

\item{land}{logical: if \code{TRUE}, plot the coastline.}

\item{ice}{logical: if \code{TRUE}, plot ice features (ice shelves, glacier tongues, and similar).}

\item{ccamlr}{logical: if \code{TRUE}, insert the CCAMLR area boundaries.}

\item{ccamlr_labels}{logical: if \code{TRUE}, add labels for the CCAMLR areas.}

\item{ssru}{logical: if \code{TRUE}, insert the CCAMLR small scale research unit boundaries.}

\item{ssru_labels}{logical: if \code{TRUE}, add labels for the CCAMLR small scale research units.}

\item{ssmu}{logical: if \code{TRUE}, insert the CCAMLR small scale management unit boundaries.}

\item{ssmu_labels}{logical: if \code{TRUE}, add labels for the CCAMLR small scale management units.}

\item{rb}{logical: if \code{TRUE}, insert the CCAMLR research block boundaries.}

\item{rb_labels}{logical: if \code{TRUE}, add labels for the CCAMLR research blocks.}

\item{sprfmorb}{logical: if \code{TRUE}, insert the SPRFMO toothfish research block boundaries.}

\item{border}{logical: if \code{TRUE}, insert longitude border.}

\item{trim}{numeric: latitude to trim the map to. Set this to -10 for effectively no trim.}

\item{graticules}{logical: if \code{TRUE}, insert a graticule grid.}

\item{eez}{logical: if \code{TRUE}, insert Exclusive Economic Zones.}

\item{eez_labels}{logical: if \code{TRUE}, add labels for the Exclusive Economic Zones.}

\item{mpa}{logical: if \code{TRUE}, insert CCAMLR Marine Protected Areas.}

\item{mpa_labels}{logical: if \code{TRUE}, add labels for the CCAMLR Marine Protected Areas.}

\item{domains}{logical: if \code{TRUE}, insert CCAMLR Marine Protected Areas planning domains.}

\item{domains_labels}{logical: if \code{TRUE}, add labels for the CCAMLR Marine Protected Area planning domains.}

\item{iwc}{logical: if \code{TRUE}, insert International Whaling Commission boundaries.}

\item{iwc_labels}{logical: if \code{TRUE}, add labels for the International Whaling Commission areas.}

\item{straight}{logical: if \code{TRUE}, leave a blank space on the side for a straight legend.}

\item{fronts}{logical or string: if \code{TRUE} or "Orsi", plot Orsi et al., (1995) ocean fronts: Subantarctic Front, Polar Front, Southern Antarctic Circumpolar Current Front. If "Park" plot the Park & Durand (2019) fronts; Northern boundary, Subantarctic Front, Polar Front, Southern Antarctic Circumpolar Current Front and Southern Boundary.}

\item{fronts_col}{character: colours to use for fronts.}

\item{land_col}{character: colour to use for coastline.}

\item{ice_col}{character: colour to use for ice features.}

\item{rb_col}{character: colour for CCAMLR research blocks.}

\item{sprfmo_col}{character: colour for SPRFMO toothfish research blocks}

\item{ccamlr_col}{character: colour for CCAMLR boundaries}

\item{ssru_col}{character: colour for CCAMLR small scale research units.}

\item{ssmu_col}{character: colour for CCAMLR small scale management units.}

\item{eez_col}{character: colour for Exclusive Economic Zone boundaries.}

\item{mpa_col}{character: colour for CCAMLR Marine Protected Areas.}

\item{border_col}{character: colours for longitude border.}

\item{graticules_col}{character: colour for graticule grid.}

\item{iwc_col}{character: colour for IWC boundaries.}

\item{domains_col}{character: colour for the CCAMLR planning domains boundaries.}
}
\value{
An object of class "SOmap", which represents a polar-stereographic map of the southern hemisphere, with the chosen management layers added. Printing or plotting this object will cause it to be displayed in the current graphics device.
}
\description{
Function for creating round Southern Ocean maps with inbuild base layers.
}
\examples{
\dontrun{
  SOmap2(ccamlr = TRUE, mpa = TRUE, trim = -45)
}
}
