% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOmap_data.R
\docType{data}
\name{SOmap_data}
\alias{SOmap_data}
\title{Contextual data for Southern Ocean maps}
\format{A list containing the following elements:
\itemize{
  \item CCAMLR_MPA
    \itemize{
      \item Description: current marine protected areas
      \item Source: CCAMLR
      \item URL: https://data.ccamlr.org/dataset/marine-protected-areas
      \item License: not specified
    }
  \item CCAMLR_research_blocks
    \itemize{
      \item Description: A defined spatial area in which research fishing on toothfish is conducted under a research plan agreed by the Commission
      \item Source: CCAMLR
      \item URL: https://data.ccamlr.org/dataset/research-blocks
      \item License: Not specified
    }
  \item CCAMLR_SSMU
    \itemize{
      \item Description: Small-scale management units (SSMUs) are designed to be used as a basis for subdividing the precautionary catch limit for krill in Subareas 48.1, 48.2, 48.3 and 48.4, and in developing management procedures for the krill fishery that can adequately account for localised effects on krill predators (SC-CAMLR-XXI, paragraphs 3.16 to 3.18). The boundaries of the SSMUs are based on predator foraging ranges (refer SC-CAMLR-XXI, Annex 4 and Trathan et al, 2008).
      \item Source: CCAMLR
      \item URL: https://data.ccamlr.org/dataset/small-scale-management-units
      \item License: Not specified
    }
  \item CCAMLR_SSRU
    \itemize{
      \item Description: Small-scale research units (SSRUs) are designed to be used as a basis for subdividing the precautionary catch limit for toothfish in exploratory fisheries, and in condcuting research fishing and developing stock assessments. The boundaries of the SSRUs are defined in Conservation Measure 41-01 (2013). CCAMLR Secretariat (2013).
      \item Source: CCAMLR
      \item URL: https://data.ccamlr.org/dataset/small-scale-research-units
      \item License: Not specified
    }
  \item CCAMLR_statistical_areas
    \itemize{
      \item Description: Statistical areas, subareas and divisions are used globally for the purpose of reporting fishery statistics. CCAMLR's Convention Area in the Southern Ocean is divided, for statistical purposes, into Area 48 (Atlantic Antarctic) between 70W and 30E, Area 58 (Indian Ocean Antarctic) between 30 and 150E, and Area 88 (Pacific Antarctic) between 150E and 70W. These areas, which are further subdivided into subareas and divisions, are managed by CCAMLR.
      \item Source: CCAMLR
      \item URL: https://data.ccamlr.org/dataset/statistical-areas-subareas-and-divisions
      \item License: Public domain
    }
  \item continent
    \itemize{
      \item Description: Coastline, details TBA
      \item Source: TBA
      \item URL: TBA
      \item License: TBA
    }
  \item EEZ
    \itemize{
      \item Description: An exclusive economic zone (EEZ) is a sea zone prescribed by the United Nations Convention on the Law of the Sea over which a state has special rights regarding the exploration and use of marine resources. Included are the EEZs that intersect the CCAMLR convention area. Some of these are currently in dispute.
      \item Source: CCAMLR
      \item URL: https://data.ccamlr.org/dataset/exclusive-economic-zones
      \item License: Not specified
    }
  \item fronts_orsi
    \itemize{
      \item Description: Southern Ocean fronts as defined by Orsi et al. 1995
      \item Source: orsifronts
      \item URL: https://github.com/AustralianAntarcticDivision/orsifronts
      \item License: see orsifronts
    }
  \item seaice_feb and seaice_oct
    \itemize{
      \item Description: median October and February sea ice extent
      \item Source: Fetterer, F., K. Knowles, W. Meier, M. Savoie, and A. K. Windnagel. 2017, updated daily. Sea Ice Index, Version 3. Boulder, Colorado USA. NSIDC: National Snow and Ice Data Center
      \item URL: https://doi.org/10.7265/N5K072F8
      \item License: Please cite
    }

   \item mirounga_leonina
    \itemize{
      \item Description: Example elephant seal Argos tracking data
      \item Source: Data were sourced from the Integrated Marine Observing System (IMOS) - IMOS is supported by the Australian Government through the National Collaborative Research Infrastructure Strategy and the Super Science Initiative.
      \item URL: https://github.com/ianjonsen/bsam
      \item License: Please cite
    }
}}
\usage{
data(SOmap_data)
}
\description{
Various spatial datasets that are commonly used on Southern Ocean maps.
}
\keyword{datasets}
